Init candidate validation tests refactor
